+++ ChangeLog.md
@@ -1,5 +1,17 @@
 ## Change Log
 
+### 1.3.6
+
+- bug fix and remove StatusBarView
+### 1.3.5
+
+- add method `hideFakeStatusBarView` to hide fake StatusBarView
+
+### 1.3.4
+### 1.3.3
+
+- fix `setColorForSwipeBack` method with `CoordinatorLayout` bug
+
 ### 1.3.2
 
 - Bug fix